pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-regblock-vhdl"
dynamic = ["version"]
requires-python = ">=3.7"
dependencies = [
    "systemrdl-compiler ~= 1.29",
    "Jinja2>=2.11",
]

authors = [
    {name="Alex Mykyta"},
]
maintainers = [
    {name="Dana Sorensen"},
]
description = "Compile SystemRDL into a VHDL control/status register (CSR) block"
readme = "README.md"
license = {text = "LGPLv3"}
keywords = [
    "SystemRDL", "PeakRDL", "CSR", "compiler", "tool", "registers", "generator",
    "VHDL", "register abstraction layer", "FPGA", "ASIC",
]
classifiers = [
    #"Development Status :: 5 - Production/Stable",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]

[project.optional-dependencies]
cli = [
    "peakrdl-cli >= 1.2.3",
]

[project.urls]
Source = "https://github.com/SystemRDL/PeakRDL-regblock-vhdl"
Tracker = "https://github.com/SystemRDL/PeakRDL-regblock-vhdl/issues"
Changelog = "https://github.com/SystemRDL/PeakRDL-regblock-vhdl/releases"
Documentation = "https://peakrdl-regblock-vhdl.readthedocs.io/"

[tool.setuptools.dynamic]
version = {attr = "peakrdl_regblock_vhdl.__about__.__version__"}

[dependency-groups]
dev = [
    "jinja2-simple-tags>=0.6.1",
    "mypy>=0.971",
    "parameterized>=0.8.1",
    "peakrdl-regblock>=1.2.0",
    "peakrdl>=1.1.0",
    "pylint>=2.6.2",
    "pytest-cov>=4.0.0",
    "pytest-parallel>=0.1.1",
    "pytest>=7.0.1",
    "wrapt>=1.14.0",
]
docs = [
    "pygments-systemrdl>=1.3.0",
    "sphinx-book-theme>=1.0.1",
    "sphinxcontrib-wavedrom>=3.0.4",
]

[project.entry-points."peakrdl.exporters"]
regblock-vhdl = "peakrdl_regblock_vhdl.__peakrdl__:Exporter"
